edit form works better now
edit form can retrieve list of saved medias
-> put method => specific javascript save action (duplicated code to
refactor later)
-> extra method on Custom\SlideShow to retrieve an array of mediaIds
(probably not the best way to deal with that, but it works)

I tried to set module path  as documented, but didn't get it to work.
Maybe some cache issue caused that, but did not want to loose too much
time on that.